Release idle thread cache when registering per-CPU.
On the slowpath, we check two conditions inside of tcmalloc::UsePerCpuCache():
* Static::CPUCacheActive()
* tcmalloc::subtle::percpu::IsFast()

IsFast() is itself composed of two pieces.  A check for IsFastNoInit and an
initialization path.

By structuring our check in terms of IsFastNoInit, we can avoid adding new
branches on the common case (using per-thread caches only / using per-CPU
caches rseq is registered) while detecting the transition from per-thread
caches to per-CPU for the thread.

This can reduce the amount of RAM pinned in vestigial per-thread caches.

PiperOrigin-RevId: 319840454
Change-Id: I4ea9dd7793c09419a28615f93d32d97e7073902c